+++ mods/Byname/Plugin.cs
@@ -26,7 +26,7 @@ namespace Byname
     {
         public const string PluginGuid = "com.ragemedia.byname";
         public const string PluginName = "Byname";
-        public const string PluginVersion = "0.3.0";
+        public const string PluginVersion = "0.3.1";
 
         internal static BynamePlugin Instance;